01-03

+++ src/main/java/com/whl/hotelService/domain/repository/BoardRepositoy.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardRepositoy extends JpaRepository<BoardEntity, Integer> {
+public interface BoardRepositoy extends JpaRepository<BoardEntity, Long> {
 
 }
